Changes -- Links (Dynamic) --- Filter Pages

+++ RentalsProject/LayerDAO/TagsDAO.cs
@@ -15,11 +15,19 @@ namespace LayerDAO
                 return db.Tags.FirstOrDefault(f => f.TagId == id);
             }
         }
-        public static List<Tag> getAll()
+        public static List<Tag> GetAll()
         {
             using (var db = new DBModel())
             {
                 return db.Tags.Include("PropertyTags").ToList();
+            }
+        }
+        public static List<KeyValuePair<Tag,int>> GetDetailAll()
+        {
+            using (var db = new DBModel())
+            {
+                var list = db.Tags.ToList();
+                return (from tag in list let sum = tag.PropertyTags.Count(p => p.PropertyDetail.isActive && p.PropertyDetail.Availablity) select new KeyValuePair<Tag, int>(tag, sum)).ToList();
             }
         }
         public static bool Delete(int id)
@@ -27,11 +35,12 @@ namespace LayerDAO
             using (var db = new DBModel())
             {
                 var tag = db.Tags.FirstOrDefault(t => t.TagId == id);
+                if (tag == null) return false;
                 try
                 {
                     db.PropertyTags.RemoveRange(tag.PropertyTags);
                     db.Tags.Remove(tag); 
-                    db.SaveChanges();
+                    db.SaveChanges(); StaticData.updateData();
                 }
                 catch (Exception e)
                 {
@@ -44,24 +53,45 @@ namespace LayerDAO
         {
             using (var db = new DBModel())
             {
+                var guid = CustomFunctions.Guid(model.TagName);
                 var tag = db.Tags.FirstOrDefault(f => f.TagId == model.TagId);
                 if (tag == null)
                 {
+                    model.guid = guid;
                     db.Tags.Add(model);
                 }
                 else
                 {
+                    tag.guid = guid;
                     tag.TagName = model.TagName;
+                    tag.Text = model.Text;
+                    tag.image = model.image;
                 }
                 try
                 {
-                    db.SaveChanges();
+                    db.SaveChanges(); StaticData.updateData();
                     return true;
                 }
                 catch (Exception e)
                 {
                     return false;
                 }
+            }
+        }
+
+        public static List<string> GetUrls()
+        {
+            using (var db = new DBModel())
+            {
+                return db.Tags.Select(t => t.guid).ToList();
+            }
+        }
+
+        public static Tag get(string token)
+        {
+            using (var db = new DBModel())
+            {
+                return db.Tags.FirstOrDefault(f => f.guid == token);
             }
         }
     }
